pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "email-triage-env"
version = "1.0.0"
description = "A real-world simulation for AI agents to manage emails."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openenv-core",
    "pydantic",
    "openai",
    "python-dotenv"
]
[project.scripts]
server = "server.app:main"